feat: unify config — env vars and config file for all settings
- Move loadConfig to @io/shared so all packages use one source
- Telegram bot token + allowed chat IDs readable from config.json
- TUI reads port from shared config (no more hardcoded default)
- Bot.ts uses config.apiPort instead of process.env
- Updated configuration docs with full reference table
- Priority: env vars > ~/.io/config.json > defaults
